Updated the Header and the cards
I added a div class to the js html string because it needed a name to help style it. I also got the css working for the cards but we need to talk in person to work on the data to loop it correctly. Right now the html loop of the js is not best apporach. Its working for tuesday talk.